py2-requirements.txt
futures
PyYAML
pyzmq
salt
six
Twisted
